separated implementation from declaration. array of jointROM became vector
